Implement the base `CompilationErrorTest`
