+++ model/build.gradle.kts
@@ -50,4 +50,6 @@ dependencies {
     testFixturesImplementation(project(":proto:configuration"))
     testFixturesImplementation(ProtoData.api)
     testFixturesImplementation(ProtoData.testlib)
+    testFixturesImplementation(Logging.testLib)?.because("We need `tapConsole`.")
+    testFixturesImplementation(Kotlin.stdLib)
 }
