+++ model/src/test/kotlin/io/spine/validation/RequiredPolicySpec.kt
@@ -26,10 +26,7 @@
 
 package io.spine.validation
 
-import com.google.protobuf.Descriptors
 import io.kotest.matchers.string.shouldContain
-import io.spine.logging.testing.tapConsole
-import io.spine.protodata.Compilation
 import io.spine.protodata.ast.Field
 import io.spine.protodata.ast.qualifiedName
 import io.spine.protodata.protobuf.field
@@ -37,65 +34,45 @@ import io.spine.validation.given.required.WithBoolField
 import io.spine.validation.given.required.WithDoubleField
 import io.spine.validation.given.required.WithIntField
 import io.spine.validation.given.required.WithSignedInt
-import java.nio.file.Path
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.api.io.TempDir
 
 @DisplayName("`RequiredPolicy` should")
-internal class RequiredPolicySpec {
+internal class RequiredPolicySpec : CompilationErrorTest() {
 
     @Test
-    fun `reject option on a boolean field`(@TempDir workingDir: Path) {
+    fun `reject option on a boolean field`() {
         val message = WithBoolField.getDescriptor()
-        val error = compile(message, workingDir)
+        val error = assertDoesNotCompile(message)
         val field = message.field("really")
         error.message shouldContain expected(field)
     }
 
     @Test
-    fun `reject option on an integer field`(@TempDir workingDir: Path) {
+    fun `reject option on an integer field`() {
         val message = WithIntField.getDescriptor()
-        val error = compile(message, workingDir)
+        val error = assertDoesNotCompile(message)
         val field = message.field("zero")
         error.message shouldContain expected(field)
     }
 
     @Test
-    fun `reject option on a signed integer field`(@TempDir workingDir: Path) {
+    fun `reject option on a signed integer field`() {
         val message = WithSignedInt.getDescriptor()
-        val error = compile(message, workingDir)
+        val error = assertDoesNotCompile(message)
         val field = message.field("signed")
         error.message shouldContain expected(field)
     }
 
     @Test
-    fun `reject option on a double field`(@TempDir workingDir: Path) {
+    fun `reject option on a double field`() {
         val message = WithDoubleField.getDescriptor()
-        val error = compile(message, workingDir)
+        val error = assertDoesNotCompile(message)
         val field = message.field("temperature")
         error.message shouldContain expected(field)
     }
-
-    /**
-     * Creates and runs a pipeline which handles only the proto type with the given [descriptor].
-     */
-    private fun compile(
-        descriptor: Descriptors.Descriptor,
-        workingDir: Path
-    ): Compilation.Error {
-        val fixture = ValidationTestFixture(descriptor, workingDir)
-        val pipeline = fixture.setup.createPipeline()
-        val error = assertThrows<Compilation.Error> {
-            // Redirect console output so that we don't print errors during the build.
-            tapConsole {
-                pipeline()
-            }
-        }
-        return error
-    }
 }
 
-private fun expected(field: Field) = "The field type `${field.type}` of `${field.qualifiedName}` " +
-        "is not supported by the `($REQUIRED)` option."
+private fun expected(field: Field) =
+    "The field type `${field.type}` of `${field.qualifiedName}` is not supported " +
+            "by the `($REQUIRED)` option."
