+++ model/src/testFixtures/kotlin/io/spine/validation/CompilationErrorTest.kt
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2025, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.validation
+
+import com.google.protobuf.Descriptors
+import io.spine.logging.testing.tapConsole
+import io.spine.protodata.Compilation
+import java.nio.file.Path
+import org.junit.jupiter.api.assertThrows
+import org.junit.jupiter.api.io.TempDir
+
+/**
+ * An abstract base for classes that test compilation errors.
+ */
+abstract class CompilationErrorTest {
+
+    @TempDir
+    private lateinit var workingDir: Path
+
+    /**
+     * Asserts that the messages represented by the given [descriptor]
+     * does not compile.
+     *
+     * It is `public`, so that extension functions could use it to create
+     * test-specific assertion methods.
+     */
+    fun assertDoesNotCompile(descriptor: Descriptors.Descriptor): Compilation.Error {
+        val fixture = ValidationTestFixture(descriptor, workingDir)
+        val pipeline = fixture.setup.createPipeline()
+        val error = assertThrows<Compilation.Error> {
+            // Redirect console output so that we don't print errors during the build.
+            tapConsole {
+                pipeline()
+            }
+        }
+        return error
+    }
+}
